Rename counters to metrics.
